Remember data from previous advertisements
Right now, as soon as one piece of data comes in, all previous ones
are forgotten. This makes it impossible to match devices by name, if
more data is received after the advertisement containing the name.

This fixes the problem on linux and windows. I've made no changes to the
mac code, as I'm not sure it even implements this.

This is lightly tested at the moment. I ran unit tests, and the linux
changes are copied by hand from a more comprehensively modified local
version, on which they work (and fix the described problem).

Windows changes are trivial but have not been tested at all.

I added a number of comments to help others trying to work with the
code. Some may be subjective, in which case, change them as desired.