AdapterManager: use broadcast channel for events
Following on from #182 this also updates the CentralEvent stream to be
based on a single broadcast channel instead of maintaining a set of mpsc
channels and tracking when to remove stale senders.

As this also related to changing shared AdapterManager state then
instead of placing the broadcast channel under a separate Arc this patch
also introduces an aggregated shared state Arc.