winrtble: union advertised service ids into a set
Previously service uuids were tracked in a vector and each time an
advertisement was received the vector would be replace - which would
tend to result in the uuids being cleared, since many advertisements
are for other things besides service uuids.

The majority of this patch is about decoupling the backend's internal
tracking of 'properties' from the public-facing PeripheralProperties
struct. This generally gives the backend more flexibility to choose
how it tracks state independent of the public API; it also enable more
fine-grained locking control over state when reading or writing changes.
In particular though it lets us use a HashSet for tracking the service
uuids which is more convenient when merging uuids.

Fixes: #189