docs: notifications() stream valid across connections
It's helpful for applications to know that they can query a
notifications stream for a Peripheral before opening any connection and
trust that the stream will remain valid across multiple re-connections
to the same Peripheral if necessary.

With the recent changes to use a broadcast channel in the winrt and
corebluetooth backends there's no longer anything special done with the
stream in relation to connections. In the bluez backend the only work
that's done is to register a device filter with dbus, which is also
independent of any connection state for the device.