winrtble: Cache is no longer used when requesting gatt services
Peripheral::connect used cache when requesting gatt services (cache is used by default if cache mode is not specified).
This means that a connection to the device might not be established if the services were previously cached.
Thus, a peripheral may not always be connected even if the `connect` method returns `Ok`.

After this change the gatt services are requested without the cache.