Prevent reference cycles in peripherals
AdapterManager and Peripherals had reference counted pointers to each other.
This prevented them from ever getting disposed. If a connected peripheral is never disconnected manually,
it would stay connected until the process is terminated.

There is now ever only a single instance of AdapterManager and peripherals have a weak reference to it.

Additionally, there was a reference cycle between Peripheral and BLEDevice in winrtble. ConnectedEventHandler in BLEDevice now uses a weak reference.