feat: Improved characteristic discovery resilience
Calling GetCharacteristicsWithCacheModeAsync() or GetCharacteristics()
would fail intermittently for services where the characteristics had
previously been discovered. This was unaffected by by the caching mode.

By gating the calls to BLEDevice::get_charactistics() in
Peripheral::discover_services(), services are only discovered once and
avoids the Access Denied errors that intermittently appear on re-discovery.

This change also causes BLEDevice::get_charactistics() to raise an error
to Peripheral::discover_services() which prevents the service from being
marked as having it's characteristics scanned in the case that there's
a problem with the discovery.